Fixes related to release building

+++ CIScripts/Deploy/Deployment.ps1
@@ -57,8 +57,10 @@ function Install-Artifacts {
         }
     }
 
-    Write-Host "Copying dlls"
-    Copy-Item -ToSession $Session -Path "dlls\*" -Destination "C:\Windows\System32\"
+    if (Test-Path "dlls") {
+        Write-Host "Copying dlls"
+        Copy-Item -ToSession $Session -Path "dlls\*" -Destination "C:\Windows\System32\"
+    }
 
     Invoke-Command -Session $Session -ScriptBlock {
         # Refresh Path
